Adds utility type class and family to aid in using type lists
This adds a type class `Elem` that checks for membership of a
given type within a type list, and a type family that can append
type lists. This will be required for the type stew that will allow
us to type-check an attribute's validity on an element at compile
time.