Adds element data tags from Blaze
This adds one data tag for every element supported by the Blaze
HTML package upon which this library will be dependent. These will
be required for the type-level element/attribute validation to
come.

Note that this DOES NOT filter out deprecated elements - that will
come at a later time, as each element/attribute is researched and
decisions are made about those associations.