Removes deprecated tags, adds documentation to those remaining
This removes the elements that are listed as deprecated by the HTML
documentation, and provides documentation explaining what each data
tag represents for those that remain.

Some elements that are listed as part of the HTML documentation are
not present in the `blaze-html` library - I may fork the repo and
open a PR to add them.