Adds Attribute validation data types
This adds a type family `ValidElementsFor`, which takes an
attribute newtype as its argument and returns a type list of all
valid element tags for that attribute. This is used by the data
type `ValidAttributeOf`, which takes an element and ensures that a
given attribute both has an instance of the `Attribute` class (a
new type class added to go from an attribute newtype wrapper to
Blaze's `Attribute` type) and is in the list of valid elements for
that attribute.

This makes it possible to enforce at the type-level (and thus at
compile time) that only attributes that are valid for a given
element can be applied to that element.

This also adds `Id`, a newtype representing the `id` global
attribute.