Adds child Element validation data types
This adds a type family `ValidChildrenFor`, which takes an element
data tag as its argument and returns a type list of element tags
representing all valid child elements for that element. This is
used by the data type `ValidChildOf`, which takes an element and
ensures that a given element is in the list of valid children for
that element.

This makes it possible to enforce at the type-level (and thus at
compile time) that only elements that are valid children for a
given element can be included as content under that element.

This also introduces a type `HTML`, which has two constructors:
`Element`, representing a fully-formed HTML element; and
`NonElement`, representing either raw text or a comment.

`Element` has a dependent type `ElementDetails` which represents
the entire element - the builder function from Blaze used to render
it, the list of (valid) attributes that will be applied to it, and
the list of (valid) children OR an indicator that it is
self-closing.

`HTML` values are built using element-specific functions that call
an appropriate `createElement` function based on whether the
element's `TagType` is `Branch` or `Leaf`.

In type signatures, `HTML` is accompanied by an element tag to
indicate what element it is - and thus, which attributes and child
elements it should have - to the compiler.

There are still some kinks to work out - currently, unwrapping the
nested child `HTML` elements from their `ValidChildOf` wrappers is
proving difficult. When it comes time to work on rendering the
HTML, I hope to have this resolved.