Adds type synonyms representing common HTML element groups
This adds a module that exports type synonyms for type lists of
element data tags. Each of these type synonyms is meant to stand
in for common element groups, either as defined in the HTML docs
or that I have found to be recurring. These are nothing more than
a convenience for writing `ValidChildrenFor` return values without
repeating ourselves constantly.

We've also removed the `Keygen` element data tag since the element
it represents, `<keygen>`, is deprecated and should not be used in
new websites.