Adds ValidChildrenFor values for most elements
This adds the return values for the `ValidChildrenFor` type family.
Most elements are covered in this, but any element with complex
conditions for what makes a valid child has been commented out and
skipped for now, along with any element that has an existing tag
group type list as its valid children with at least one exception.