Multiple additional failed, plus one working, attempt
The previous attempts tried different approaches around how to
define attributes and child elements, but all of them ultimately
ran into Haskell's distaste for heterogeneous types in a way that
could not be thought around - at least by me.

`HTML4`, however, shows promise - instead of trying to define what
child elements are valid for an HTML element, we're instead now
defining a category, with the element constructor  asking for
(in addition to its valid attributes) a category to which its
child elements can belong, and is returning a `Node` defined by
what category it belongs to. We then hide this with the `HTML`
type synonym which takes the type of element it's defining and
uses a type family `ElementCategory` to determine which category
the element should be in.

The next step is going to be figuring out how to represent the
rather complicated rules around what categories to which an element
may belong, and how to get `Node` to enforce it.