Adds type synonyms for TagType constructors, further experimentation
This introduces a `TagTypes` module to serve the same purpose as
the old tag modules, though rather than data tags, these are type
synonyms for the `TagType` constructors to represent in the GADTs
of the `Elements` modules. Some further experimentation is done on
`Attribute` and `HTML` to see what is possible with the types in
their current state. The next PR will be laying out simple code for
rendering `HTML`.