Adds basic code to render HTML as Text
This adds the fundamentals that will permit rendering to `Text`. It
is this basic blueprint for rendering that will be copied to the
`String` and `ByteString` rendering functions, eventually.

Currently, the rendering is minimized. There is no pretty renderer
as of yet - pretty rendering is not needed for my purposes, and to
implement a pretty renderer with HTML (where whitespace can change
the result of the rendering) has a lot more considerations than I
would be willing to put in right now while I'm still not sure that
this iteration of the library will be the one I move forward with.