Adds working example of HTML checking children per-element
This modifies the previous approach to contextualize `HTML` (now
called `ChildHTML`) in terms of its parent element, not in terms
of its content category. This allows us to check a type list of
elements to ensure that all of an element's content is valid as a
child of that element, while still monomorphizing the list of
children in a way that allows it to contain different elements.

Also in this commit: `renderHTML` the conditions backward for
determining when to put a space after a tag name (when there are
attributes to follow).