Some organizational and documentation changes
This moves the `ValidChild` type synonym and `ValidChildrenFor`
type family to a new module specifically dealing with declaring
valid children for an HTML element.

This also uses the `Tags` module type synonyms everywhere that may
end up being user-facing, since those are what will ultimately be
exposed for their user.

Finally, it updates the documentation on those `Tag` type synonyms
to remove any `@HTML@` references since the name has changed and
referring to it as `ChildHTML`, especially in this module since it
doesn't import that type, seems misleading. Instead, the docs now
simply state that a given type synonym represents the appropriate
elements or HTML entity (such as text or a comment).