Adds all ChildHTML constructors
This adds all the constructors for the `ChildHTML` GADT, but does
not necessarily set them correctly - some elements that cannot
contain child elements are still expecting a content argument.
This was mostly a mass copy with corrections to bring the types in
line with the HTML spec still to come.