Adds all straightforward ValidChildrenFor instances
This includes all of the ValidChildrenFor instances that are
reasonably straightforward to implement - a short list of specific
elements, an entire content category, or some minor departure from
those. Left out are those with conditional permitted content lists
and those that are an entire content category with some exceptions.
These will be added in the future.