Adds element builder functions
This adds a builder function for every constructor of `ChildHTML`,
as well as a few convenience functions. There is one commented out
function `table'` which is meant to be a safer version of `table`,
but I haven't yet managed to get it to compile. I'll return to it
in the future.

This was built using code generation, which I haven't saved since
it won't be required again.