Adds Table module for safe construction of tables
This adds a new module `Data.HTML4.Elements.Table` that safely
constructs an HTML table with the child content in the correct
order as per the HTML spec. It does not ensure that each row has
the same number of columns, since that can't be done at the type
level. Since we can view the presence of a `colspan` attribute on
elements, perhaps a function to add this additional layer of safety
could be added as an even safer function in the future.