Makes comments valid everywhere, adds some convenience functions
This removes the `ValidChild` constraint for HTML comments at the
constructor level, making them viable on all elements that can
hold children.

Additionally, this adds a few convenience functions:
  - `noElement`, which renders an empty string and, like comments,
is viable on all elements that can hold children. This is useful
in `if` expressions and on `Maybe`s when you only want to render a
`Just`.
  - `wbrs`, which takes a list of texts and puts a word break
opportunity between each of them. This is a convenience for users
to break a string where they know it makes sense without having to
put in the `wbr` tag themselves.