Adds transparent tag group, updates transparent tag children
For tags that purely contain transparent children (`ins`, `del`,
and `slot`), this modifies their content argument in the `ChildHTML`
constructor to take a list of children for their parent element.
Their instances of `ValidChildrenFor` have been removed.

Additionally, since `map` can take any transparent element, this
adds a tag group for transparent elements and uses that as the
argument for the `ValidChildrenFor Map` instance.