Adds ByteString rendering
This adds almost an exact duplicate of the existing `Text` render
function, but instead it renders a lazy `ByteString`. This is what
is used in the body of HTTP responses using the `http-conduit`
library, so it is likely the render function to see the most use
besides `Text`.

A third function may come about eventually for `String`.