Adds String rendering
This adds almost an exact duplicate of the existing `Text` render
function, but instead it renders a `String`. The to-text functions
for supporting types will receive a to-bytes version in the next
commit to squeeze more performance out of the bytestring render
since it's the primary one I'll be using for SSR, but this one
will not receive similar treatment.