Adds to-bytes conversion functions
This represents a compromise between speed and ease of use. Lazy
`ByteString` is likely going to be the most used render target
since it's the type used in HTTP response bodies in the Haskell
HTTP library, so having to-bytes functions instead of converting
after the to-text functions will be more efficient. However, I
stopped short of having the types carry bytestrings instead of
text because text is easier to work with than bytestring.