Adds RawHTML constructor for ChildHTML
This adds a new constructor storing `Text`. `RawHTML` is different
from `Text` in that is rendered unescaped. This allows users to
insert raw text that we can't do anything with, like templating
engine commands. It's also intended to be used for writing the
content of `<script>` and `<style>` tags. However, because it is
unchecked and unescaped content that is valid anywhere, it is
documented as unsafe. It is used at the user's own risk.