Adds CustomHTML constructor for ChildHTML
This adds a new constructor `CustomHTML` that allows users to
define their own HTML element. By its very nature, this element
cannot be type-checked for validity in relation to any element; it
is both considered a valid element for all possible parents, and
considered to be a valid parent for all elements.