Adds more tag operations, valid child definitions
This adds more `ValidChildrenFor` instances, specifically covering
all elements except for `<noscript>` and elements with complex
conditions around transparency.

Most of the elements that received valid child definitions in this
commit accepted one or more content groups, but had some manner of
exception regarding specific excluded elements. To simplify the
definitions, the contents of `HTML.Contains` was moved to a new
module `HTML.Internal.TagOperations`, which now houses all of the
type class and type family definitions that deal with operating on
`TagType` lists to discover or revoke membership in said lists.

Unfortunately, nested type family applications could not be avoided
here, so I've made the decision to turn on `UndecidableInstances`.
I have deemed it safe enough to avoid changing the entire approach
to the library at large, given that we can be fairly certain that
all of our type family definitions have a base case defined and
that all arguments passed to them are finite lists that will
eventually reach said base case(s). I'm happy to accept any
proposed alternatives that would avoid this without significantly
changing the API or negatively impacting the user experience. As
of now, the current approach hasn't changed the fact that elements
type-check at compile time and that the user doesn't have to turn
on any of these iffy language extensions themselves in order to
interact with the API - I would like to keep it this way.