Changes script and style to take `Text` directly
Both of these can only have raw text content, whether it represents
a style sheet or a dynamic script. In either case, taking a list
of `ChildHTML` that is limited only to `RawHTML` seems silly since
it can only ever be a single element long anyway. Instead, we just
take `Text` now and convert it into a singleton of `Raw_HTML` when
we render.