Adds custom error messages for ValidAttribute and ValidChild
This moves away from the `Contains` type class in favor of three
new type families. `Elem` checks for a tag's membership in a list
of tag types, while attributes and elements each report a custom
compiler error message when their respective `Alert_` type
family determines that the tag is not a member of the tag type
list.

Also in this commit is a fix a typo in what was `FormOnValidate`
and is now `FormNoValidate`.