Makes Object a transparent element
Object is transparent except for also accepting param elements,
which are deprecated and thus not going to be supported by this
library. This means object can be treated as a completely
transparent element, like ins and del. This is what was done here.

This also does away with the idea of the `FlowContentBase` tag
group and just adds header and footer back into the flow content
tag group.