Adds support for complex transparent elements
With the grandparent param for `ChildHTML` in place, we are able
to create a complete list of tag types to check built from the
specific permitted and excluded transparent content and the list of
valid elements for the grandparent (that is passing those valid
children along to the parent).

This also fixes the type checking on the content argument for all
of the `ChildHTML` constructors and element builder functions by
setting their grandparent to the current tag's parent, rather than
what would be the content element's great-grandparent.

The only remaining transparent element (and the only remaining
element that doesn't have a complete `ValidChildrenFor` definition)
is `NoScript`, which is complex. It will be added in a future
commit.