Adds valid children cases for NoScript element
`NoScript` has different valid children depending on what its
parent is. This adds the `ValidChildrenFor` instances necessary to
represent this case - that is, for the most part. This merges two
cases into one: when scripting is disabled and it's a descendent of
`Head`, and when scripting is not disabled. We simply cannot
determine if scripting will be supported or turned on in the
browser at this level, so we provide all the options possible for
the API consumer as the base.