Removes `Document` data type
This was originally added with the thought that this system could
support XML, so specifying what kind of document was important.
This ends up being somewhat frustrating to use in practice because
different functions have to be written to send either a complete
HTML document or HTML snippets. Since the only actual usage of
this library currently is for an HTMX application, making the
latter easier is the priority.

That said, adding an `XML` constructor to `ChildHTML` is not out
of the question, but the only thing it could take would be custom
attributes. I can add this down the line, at which point I might
change the name of `ChildHTML` to `ChildElement` or `ChildTag` to
reflect its multi-purpose nature.