Adds RelativeURL type
This type leverages the Beeline `RouteGenerator` to go from a route
definition to a method and rendered path. Because we're using the
routers from Beeline, the route can reflect a route on the server's
existing router, so it will be flagged by the compiler in the event
of a change. This means the relative URLs defined in `href`
attributes won't suddenly break if there's a change to the router.