Adds Href type
This has been implemented as a Shrubbery union so that types can
be passed to a builder `href` function that does the unification.
This helper, along with the more complex logic around what types of
href can be accepted for a given element, will be implemented in a
later commit.