Adds Attr_Href Attribute constructor and builder function
This adds support for the `href` attribute. The builder function
`href` uses a new type family `ValidHref` to ensure that whatever
type is being passed to the `mkHref` function (which must be a
member of `HrefTypes` from the previous commit) is valid for
whatever tag it's being added to. This new constraint makes use
of a custom error message when it comes up `False`.

The `href` attribute is valid on `a`, `area`, `base`, and `link`,
but the latter two are only able to accept absolute and relative
URLs, where the first two can take any kind of href. This was the
reason behind the type family restriction.