Gives RelativeURL a method context, blocks non-GET on href
In addition to the limitations on what type of link `href` can
accept based on element, `href` realistically is only able to hold
routes that represent HTTP GET requests for an absolute or relative
URL. We can't guarantee what kind of route an `AbsoluteURL` will be
since it's external, but because we make `RelativeURL`s from
Beeline routes, we can enforce the method used here.

By adding this context to `RelativeURL`, we can further restrict
the `ValidHrefsFor` result to accept NO valid tags for anything
other than an HTTP GET. When we get around to making the `action`
form attribute, we can restrict it to HTTP POST routes as well.

This is done by replacing the Beeline functions to create a route
with custom ones that provide a narrower list of methods. HTMX can
only support GET, POST, DELETE, PUT, and PATCH; normal HTML
attributes support even fewer. Beeline, however, supports all nine
HTTP methods. This is fine for endpoints, but not for routes we're
converting to `RelativeURL`, so we use custom functions for each of
the five methods we want to support, and then build them at the
end using the Beeline equivalent. The path is now the only thing
being held by the `RelativeURL` since its constructors, once cased
upon, tell us the rest.