Adds HTMX attribute `hx-ext`
This was done by introducing a new type `Extension` that holds an
`ExtensionType` and a `Bool`, indicating whether it should be
ignored. These type constructors aren't exposed, so instead users
call helper functions that default the extension to be not ignored,
with another helper function `ignore` to deliberately flip the
extension to be ignored.