Exports Id constructor, makes Class a newtype
This gives `Class` the same treatment as `Id` got, where it is now
a newtype wrapper around `Text` so it can be more easily shared
with other code that takes `Class` values, like some upcoming HTMX
attributes.

The `Id` constructor is also exported for ease of use. It can now
be used to build and destructure `Id` values, but `idFromText` is
still exported for easier function composition.