Adds `ClassSelector` type for upcoming query selector feature
This adds a `ClassSelector` type, which holds a `Class` and can be
negated (ex: `div:not(.myClass)`) using a type class `Negatable`.

This also adds `_toBytes` functions for both `Id` and `Class`,
though more will have to be done to figure out how and when is best
to escape attribute values before these will enter wider use.