Adds `AttributeSelector` type in preparation for query selector
This adds a new type `AttributeSelector` which represents an
element's attribute and value to match on in a query selector.

Unfortunately, we cannot rely on the existing type safety measures
for attributes (or elements, for that matter) in query selectors,
but I think it's okay to accept this limitation given that a query
selector can also work on elements that aren't necessarily valid
combinations, and because it may have to select on HTML that wasn't
generated with this library, so might be invalid anyway.

For ease of use, this adds a builder function for each attribute
(again). Currently, they're all simply taking text - this is a
shortcoming that was chosen for time and will be rectified over
time. The new process for adding an attribute will include ensuring
that these attribute functions are taking and converting the
appropriate types.

This also fixes a few attribute-related bugs, including a bad
render on an HTMX attribute and the removal of a nonexistent
"buffer" attribute.