Adds `ElementSelector` type in preparation for query selectors
This adds an `ElementSelector` type that will serve as one of the
top-level options for a future query selector type. This accepts
the previously-added `AttributeSelector` and `ClassSelector` types
in a recursive structure to create complex query selectors on
elements.

Like `AttributeSelector`, this adds convenience functions to build
`ElementSelector`s. These look similar to the element builder
functions in the main portion of the library. Some name collisions
forced the addition of a "Tag" suffix to some elements.

Also like the `AttributeSelector`, the spec-enforcement mechanisms
in place for actual elements and attributes cannot work here since
they would create cyclical dependencies. To provide at least some
degree of enforcement of HTML spec rules here, elements that cannot
hold children, or elements that only accept text content, will not
permit nesting.