Adds `QuerySelector` type
This adds a `QuerySelector` type that can be used in conjunction
with upcoming HTMX attributes such as `hx-target` and `hx-swap`.
It's using the union construction that's becoming standard for this
library, and can accept any of an id, class, or a complex query
selector targeting nested elements optionally by their attribute
values and classes.

This also fixed a bug found during testing where the rendering of
`ClassSelector` was flipped - it would render negated classes as if
they were desired and vice-versa.