Updates `AttributeSelector` builders with appropriate types
With the exception of `href`, all currently supported attribute
builders has the types of its arguments reflected in its attribute
selector counterpart.

`href` is an exception because of some issues with the type soup
required to get it to accept one of a number of options in the
union. One possible solution to this is to introduce a secondary
union just for this, which works mostly the same way.