Adds HTMX attribute `hx-on`
This uses the `EventType`s introduced in the previous commit to
create another union `Event`, which accepts any of the three
defined event types.

Also worth noting is the conversion of the `configSchema` from a
generic schema type to an `Encoder` - this will only ever be used
to encode to JSON, never to decode. This distinction was worth
making here since it relates to an issue around attempting to
decode `EventType`s since a handful share namespaces in the HTML
events list.