Removes Negatable type class
This was originally done to accomplish a more aesthetically
pleasing way of writing HTMX code that would more closely reflect
how it looks in raw HTML, but using a function to accomplish the
same thing is servicable. Unless it turns out that more attribute
argument types are also "negatable," this class (and likely, the
approach it represented) will not reappear.