Adds None type, renames other types, removes RawTarget
This adds a `None` type that will be used in the upcoming `hx-swap`
implementation. To avoid name collisions, the constructor sharing
that name for `RequestParams` was renamed to `NoParams`.

Additionally, `RawQuerySelector` was renamed to `RawSelector` so
it could be reused with other selector-affiliated types that can
accept raw text content. It was then made to replace `RawTarget`
since it serves the same purpose.