Adds HTMX attribute `hx-select-oob`
This one is reusing the `SwapSelector` from the previous commit as
part of its union, as well as `QuerySelector` and `RawSelector`.
One notable part of this is that, because it can accept a list
(nonempty) of these unions, they have to be wrapped ahead of time,
so they can't be used quasi-dynamically like the other union types.