Adds remaining union types for `hx-swap`
This adds the scroll and show behavior modifiers for `hx-swap`.
This is another data type that holds a union on one of its fields,
and constructs it with the same quasi-dynamic interface as most
other unions.

This does NOT implement the attribute itself - that is saved for a
future commit.