Adds (very lean) support for the lang attribute
`lang` uses the BCP-47 standard for language codes, which is way
more complicated than is reasonable to do in order to support such
a common attribute. Until the need arises for a more complex
implementation (or there is nothing else more important to do),
we'll settle on this extremely straightforward approach.