Adds charset, content, and name attributes
`charset` is fairly straightforward - it only supports UTF-8
encoding as that's the only valid option for HTML 5.

`name` and `content` are a bit different. They are inter-dependent,
with the content of one depending on the content of another. To
complicate matters further, this is only true within the `meta`
tag - other uses of `name` do not have this restriction.

Therefore, we support them in the base `Attributes` directory as
accepting a simple `Text`. This should service most of the use of
`name` in form fields. To support it safely in `meta` tags, it's
a mandatory argument in `Brigid.HTML.Elements.Safe.Meta`, where we
represent it as an ADT with each constructor taking a structured
data type of some sort that properly represents its valid values.