Adds beginnings of `Safe.Script`
This breaks the `script` tag down into its two fundamental builds:
inline and external scripts. A `Shrubbery.Union` is used to
represent these options, with each containing details about the
attributes (and a content field for the former). Since they were
straightforward to implement, the following attribute combinators
have been created as well:

- `async`
- `defer`
- `referrerpolicy`

The last of these was lifted from `Safe.Meta` so it could be used
for both.