Adds HTML entities module built from web scraper script
This adds a new script `fetch_entities` which runs a Python module
to fetch HTML entities from a web source, scrape the table with
HTML entity data, and build a Haskell module for Brigid with a
combinator function and explicit export for each HTML entity. This
script is now run as part of the build script.

As for the combinators themselves, they are using a new `ChildHTML`
constructor `Tag_Entity` which has the same constraints as
`Tag_Text`, but does not do any HTML escaping when rendering.