Removes `addAttribute` again
This was an unintended consequence of the commit reversion as
this was done alongside the removal of the `Map` structure for
attributes. We want to keep the `Map` structure for the free
deduping, but remove `addAttribute` as we no longer intend to
encourage this sort of approach (which can be done using `cons` on
an attribute list anyway).