Removes OMap, replaces with nubOrdOn
This removes the use of `OMap` to dedupe attributes and instead
uses `Data.Containers.ListUtils.nubOrdOn`, which uses the same
`attributeText` function that was used as the key values in the
`OMap` to do the deduping.

This simplifies a lot of the code, and should be the most
performant compared to having to fold over the attributes (`O(n)`)
in order to insert (`O(n log n)`) into the `Map`s, then fold over
them again (`O(n)`) to get the `elems`. This is just one single
pass over the list with a function with `O(n log n)` time
complexity.