Adds BCP_47 type alias for ISO639_1
This adds a new module to handle the planned type `BCP_47`. This
will be a more robust representation of the BCP-47 standard for
language codes in HTML. For now, it's just a type alias for
`Data.LanguageCodes.ISO639_1` so as to affect minimal change.
This is done because support for an upcoming attribute will also
require BCP-47 values.