Adds custom fold functions for rendering attribute value lists
Some attributes hold collections of values rather than a single
value. When rendering these values previously, either for actual
rendering of the HTML or for the `QuerySelector` module, we would
`fmap` over the collection to convert it to `ByteString` or `Text`,
then often we would `intercalate` (or `unwords`) the list. This
results in multiple passes over the list.

To make this rendering more efficient, we now instead fold over
them once and make all the desired changes in a single pass with
custom fold functions. These functions take the separator as an
argument so we can use it replace `intercalate` and `unwords` at
the same time.

These functions are then used everywhere we render an attribute
that stores a collection of values.