Refactors `TargetType`, adds `Target` ADT
This adds a new ADT for values that will eventually be used in the
`target` attribute. Because this namespace was taken for values in
use with the `hx-target` attribute, that type has been renamed to
`TargetType`. Additionally, constructors for `SwapDisplayView`
had to be changed to avoid name collisions with the `Top`
constructor of `Target`.