Adds support for the datetime attribute
This also adds a function `datetimeWithFormat` to allow users to
enter their own format string, at their own risk. This allows
users to use format strings for week number or year-month formats,
which are permitted by ISO 8601 but not accessible through the
`Data.Time.Format.ISO8601` module. It's more flexible than HTML
actually allows, but a warning is provided in the documentation to
exercise caution when using this looser function.