Minor refactor to existing types in preparation for safe input tag
This renames `HTMLEvent` and `KeyHintOption` constructors to avoid
name collisions with the upcoming input type ADT constructors we
need to support a safe module for the input tag and the `type`
attribute.

Also done is a slight refactor to the `URL` types module - namely,
exposing a function to make a `URL` instead of its constructor and
exports `rawURLToBytes`.