Adds type option union for handling valid type attribute values
The `type` attribute is one of the more complicated ones in the
whole of the HTML spec. Depending on which of the many children for
which it's valid, the values it's expected to hold would be best
represented as different types entirely. This union allows for the
combined representation of `type` attribute values in one type,
and uses another type family to ensure only valid `type` options
make it into the attribute for a given element, like the `rel` and
`src` attributes.