Adds Safe module for input tag
This tag exports type-specific combinator functions for each of
the valid `type` attribute values on the `input` tag. New tag
groups were added, and existing ones were modified, to support the
new input tags for existing scoped attributes. Additionally, the
output text of `TagErrorMessage` was corrected for the input tags.