Adds dependency on Ogma
Ogma is a new library for handling languages - ISO 639 codes,
ISO 15924 codes, and primarily BCP-47 codes. The goal is that it
will be able to leverage Haskell's type system to provide a
similar service to Brigid with its HTML spec compliance at compile
time, but with BCP-47 codes, which can be quite complex and for
which no satisfactory (for the author, anyway) library for their
safe creation exists.