Adds Value union for handling possible options in value attribute
Much like `Relationship` and similar types before it, `Values` is
a `Shrubbery.Union` with its own type family to ensure that the
value being added to the upcoming `value` attribute is valid for
the tag it's being placed on. Currently, it's incomplete - values
still need to be added for the `InputNumber` and `InputRange` tags,
and we still need to verify what types of `RelativeURL` are valid
in the `value` field for `InputUrl` tags. These are lower priority
and can be completed at a later time.