Adds support for the value attribute
This is another attribute like `rel` that has a limited selection
of permissible values depending on what element it's been placed
on. This is handled with the `ValidValue` constraint type family
that was added in a previous commit. Many of its permitted values
are types that were also added in previous commits.