Adds support for the max and min attributes
To support this, we've also added a new `RangeBound` union type so
these complex attributes can support the possible values for the
varied `input` types that use it. All of these values have been
given `Ord` instances so we can support sorting the values and
ensuring the min/max are appropriately set in the `Safe` modules
in the future.