Updates name attribute to take a union type
This `NameOption` union type allows for both the ADT-based options
for the `meta` tag and a free-text `Name` newtype for all other
elements that accept the `name` attribute. A new type family
`ValidName` ensures that `MetadataName` is restricted only to the
`meta` tag. This means we can now use the `Name` newtype for
attributes like `usemap`.