Exposes common text newtype constructors
This exposes the constructors for `Id`, `Class`, and `Name`, and
removes the `mk_` functions. It is likely that the underlying text
for these, especially for id and name, will be shared on a tag, so
by exposing the constructor we make them `Coercible`.

Also done: we switched to the `custom` attribute rather than the
`customAttribute` function when setting the HTMX config. We still
need the `customAttribute` function for defining the name since we
don't want to add HTMX functionality in the base web section. We
instead want to continue providing `setConfig` to allow HTMX users
to create the appropriate `meta` tag.