Adds AutocompleteToken type
This union type will be used with the `autocomplete` attribute. In
order to support this in a way that satisfies the HTML spec, we
had to create it as a union holding simple data tags for every
possible autocomplete token value. This is because each token value
has different rules about which element it can belong to, as well
as which can be used with token modifiers such as `section` or
`home`.