Adds Modifier GADT, removes other modifier types
This adds a GADT `Modifier token`, which allows for constraints
specific to the constructor to be enforced and therefore allow the
smart constructor functions for each modifier to impose the validity
constraint type classes. This allows us to ensure that only valid
autocomplete tokens end up on the right modifiers, and (eventually)
that they end up only on valid elements.