Adds support for the http-equiv attribute
This also adds a very basic representation of this attribute in the
`Safe.Meta` module. Currently the ADTs in both places (the safe and
`HTML.Types` modules) are using the exact same constructors, but
eventually the `Safe` version will be expanded upon to include the
valid `content` attribute values - this is why the ADT doesn't have
`Bounded` or `Enum` instances in the `Safe` module.

Instances of `Bounded`, `Enum`, `Eq`, and `Show` were added to
other types as appropriate.