Adds support for the pattern attribute
Unless we were to become dependent on a regex library, we could
never validate that the value passed to this attribute combinator
is a valid JavaScript regex. Even if we took that route, we could
never guarantee that this attribute combinator would produce a
safe result, because not all browsers support all features of
JavaScript regular expressions. So, for simplicity's sake, we opted
for a simple `Text` argument. We may change this in the future,
but it would have multiple smart constructors to guard against
certain JavaScript features that the user may know they don't
intend to support.