Adds support for the srcdoc attribute
This attribute takes literally an entire rendered HTML document
(or at least any valid HTML, even if just a single element) as its
argument. To accomplish this, it takes `ChildHTML parent grandparent`
as its argument. Because this make contain attributes, we have to
escape the content to avoid printing a `"` character. This meant
adding to-bytestring escapement functions, which hadn't existed
prior to this since all string-based processing favored `Text`
due to it being the preferred choice for user input. These new
functions were created and their use propagated to both render
modules.

I honestly still cannot believe that this works.