Adds support for the inputmode attribute
While this is a global attribute, it's valid all the time for only
`input` (except for the `hidden` type) and `textarea` elements.
However, it's valid on any element with the `contenteditable` mode,
but there's no way to guard against that for any attribute besides
smart constructors in the `Safe` module, which we may do. Either way,
putting this attribute on something without that turned on doesn't
hurt anything, so there's no reason to continue blocking on it.