Moves internal render module to a shared location
This moves `Brigid.HTML.Internal.Render` to `Brigid.Internal.Render`
so it can be used by the HXML and XML modules. Nothing about this
module is unique to the HTML module, and importing a module with an
HTML namespace into the HXML and XML modules will be confusing.

Also done is the curation of exposed modules to only the re-export
modules we actually want exposed.